Default enemy and enemy manager
+Weapon collider fix

+Appearing objects system

+A* pathfinding asset

+Default enemy
+Enemy manager system

+++ Assets/Resources/Scripts/Player/Weapon/LaserShootVariation.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class LaserShootVariation : ShootVariation
 {
-    [SerializeField] private float AdditionalDamage;
+    [SerializeField] private int AdditionalDamage;
     [SerializeField] private float AdditionalShootDelay;
     [SerializeField] private float TimeBeforeShoot;
